Create The Smallest and The Greatest Strings
Take as input a number N. N lines follow with one string on each line. Your task is simple, print two lines. On the first line print the lexicographically smallest string of the N strings. On the second line print the lexicographically largest string of the N strings. 

Sample Input 0  :

4  

the lord of the rings  
the godfather  
the before trilogy  
rocky balboa 

 

Sample Output 0  :

rocky balboa  
The Lord of the Rings