Create MaxElement
Instructions:
Maximum Element of Binary Tree: (Binary Tree) 

Problem Statement:  Given a Binary Tree, Print the value of maximum node key. 

Argument/Return Type 

Input of total no.of nodes is taken 

Input of key values of nodes of tree are taken in level order form  

Incase of a null node , -1 is taken as input 

A function which returns the maximum element  

If the root is NULL it prints -1 

Solution: 

if node is NULL , -1 is entered as it's key 

Sample Test Case:   

Input Binary Tree :             

 

Input:  

Enter total no.of nodes of the input Tree ( including NULL nodes ) : 15 

Enter value of each node of the tree in level order ( if a node is NULL , enter -1 ) with spaces 

10 11 12 5 -1 6 13 4 -1 -1 -1 8 9 10 -1 

Output: 

The maximum value of the node key of given tree is 13.