Create BST to Heap
Given a Binary Search Tree. Convert a given BST into a Special Max Heap with the condition that all the values in the left subtree of a node should be less than all the values in the right subtree of the node. This condition is applied on all the nodes in the so converted Max Heap.

Example 1:

Input :

4

/ \

2 6

/ \ / \

1 3 5 7

Output: 1 2 3 4 5 6 7

Explanation:

7

/ \

3 6

/ \ / \

1 2 4 5

The given BST has been transformed into a

Max Heap and it's postorder traversal is

1 2 3 4 5 6 7.

Your task:

You don't need to read input or print anything.

Input: root of the tree

Action: BST to max heap.

A method is required to print the post order traversal of the converted Binary search tree.